third_party/WebKit/LayoutTests/virtual/origin-trials-features-disabled/http/tests/origin_trials/webexposed/longtask-origin-trial-interfaces-expected.txt

CONSOLE MESSAGE: line 15: Interfaces in document
interface PerformanceLongTaskTiming
    getter attribution
    method constructor
This is a testharness.js-based test.
PASS LongTaskObserver related interfaces in Origin-Trial enabled document. 
Harness: the test ran to completion.
